wireless-regdb: Update regulatory info for Mexico (MX) on 6GHz
The Plenary Session of the Federal Telecommunications Institute (IFT)
determined, in its IV ordinary session of 2023, to classify the
5925 - 6425 MHz frequency segment as free spectrum and issued the
technical operating conditions that allow coexistence with the existing
services in the country [2], which [1] is the entry page.

Within the operating conditions of this band as free spectrum, its use was
determined for low power indoor and very low power indoor and outdoor
systems.

In the agreement [3] of section Single Annex 2. Low power systems operating
indoor is
	30 dBm (5 dBm over any 1 MHz bandwidth) for Access points
and
	24 dBm (-1 dBm over any 1 MHz bandwidth) for Client computer

Very low power systems operating indoors and outdoors is
	14 dBm (1 dBm over any 1 MHz bandwidth) for User terminal

Considering -1 dBm/MHz for Client computer and smallest channel is 20 MHz,
limit 12 dBm is adopted.

[1] https://www.ift.org.mx/comunicacion-y-medios/comunicados-ift/es/el-pleno-del-ift-clasifica-la-banda-de-5925-6425-mhz-como-espectro-libre-comunicado-132023-16-de
[2] https://www.ift.org.mx/sites/default/files/comunicacion-y-medios/comunicados-ift/comunicado13ift_1.pdf
[3] https://www.dof.gob.mx/nota_detalle.php?codigo=5681829&fecha=07/03/2023#gsc.tab=0

Signed-off-by: Ping-Ke Shih <pkshih@realtek.com>
Link: https://lore.kernel.org/r/20240605074209.8713-4-pkshih@realtek.com
Signed-off-by: Chen-Yu Tsai <wens@kernel.org>