wireless-regdb: Update regulatory info for Argentina (AR) on 6GHz
NATIONAL COMMUNICATIONS ENTITY issued Technical Characteristics WiFI6
Equipment on 05/22/2023 by Appendix 1 of [1], which 1) an access point or
subordinate device may operate in the 5925-7125 MHz band with a Power
Maximum Radiated Isotropic Equivalent (EIRP) of 30 dBm and a maximum
EIRP spectral density of 5 dBm in a bandwidth of 1 MHz; 2) A client device
may operate in the 5925-7125 MHz band with an Isotropic Radiated Power
equivalent (EIRP) of 24 dBm and a maximum EIRP spectral density of -1 dBm
over a width of 1 MHz band.

Considering -1 dBm/MHz for client device, 12 dBm is adopted.

[1] https://www.boletinoficial.gob.ar/detalleAviso/primera/287126/20230524

Signed-off-by: Ping-Ke Shih <pkshih@realtek.com>
Link: https://lore.kernel.org/r/20240621080348.55504-1-pkshih@realtek.com
Signed-off-by: Chen-Yu Tsai <wens@kernel.org>