wireless-regdb: Update regulatory info for Azerbaijan (AZ) on 6GHz for 2024
The ICTA established under the Ministry of Digital Development and
Transport of the Republic of Azerbaijan announced in June-September,
the State Commission for Radio Frequencies responded positively to 17
applications.

At the meeting of the State Commission, taking into account the
Decision No. (20) 01 of the Electronic Communications Committee of the
Pan-European Conference of Postal and Telecommunications Administrations
(ECC Decision (20) 01) and international experience, the "Conditions for
the use of the 5945-6425 MHz radio frequency band for Wireless Access
Systems (WAS), including Radio Local Area Networks (RLAN)" were approved.

ECC/DEC(20)01 defines:

* 5945-6425 MHz
  - LPI, 23 dBm, indoor use only (adopted)
  - VLP, 14 dBm, indoor/outdoor

[1] https://icta.az/show-media-news/iyun-sentyabr-aylarinda-radiotezlikler-uzre-dovlet-komissiyasi-17-muracieti-musbet-cavablandirib?

Signed-off-by: Ping-Ke Shih <pkshih@realtek.com>
Link: https://lore.kernel.org/r/20241127093208.4693-1-pkshih@gmail.com
Signed-off-by: Chen-Yu Tsai <wens@kernel.org>