wireless-regdb: Update regulatory rules for Austria (AT)
According to the Austrian "Frequenznutzungsverordnung" Attachement 2
page 404
(https://www.ris.bka.gv.at/Dokumente/Bundesnormen/NOR40251378/II_61_2023_Anlage_2.pdf),
Austria extended the usable range according to the EU SRD Band 75a (Implementing
decision 2019/1345, current consolidated version:
https://eur-lex.europa.eu/legal-content/EN/TXT/?uri=CELEX%3A02006D0771%2801%29-20220210&qid=1738145465756)

This probably applies to most other countries that are part of CEPT
and/or EU but I'm not good enough with international law to be sure
about others.
It might be worth checking and updating this for many more countries.

This also updates two outdated links (the RTR one was a 404, the CEPT
one redirects with a warning that this link wil stop working).
Specifically the CEPT one might need updating for all other countries as
well (also the link now seems to refer to a newer version of the document
than the originally referenced one. v2.2.1 is from 2021 and was updated
to include the 66-71GHz range as well).

Signed-off-by: Jakob Riepler <jakob.riepler@paranoidlabs.org>
Link: https://lore.kernel.org/r/20250129112554.27296-2-jakob.riepler@paranoidlabs.org
Signed-off-by: Chen-Yu Tsai <wens@kernel.org>